requirements.txt
streamlit==1.31.1
Pillow==10.2.0
pytesseract==0.3.10
google-generativeai==0.3.2
python-dotenv==1.0.1 